Return NullAdapter when calling dotted properties
Now returning a NullAdapter if a call with a dotted path catches a
property in the chain while there are still properties to be
resolved. This avoids returning a value for 'foo.bar.baz' if
'bar.baz' is an invalid path but 'foo' exists as property. Previous
the call for 'foo.bar.baz' returned the value for 'foo'. Now this
scenario returns a NullAdapter since 'bar.baz' is not a valid path.